+++ crates/threshold-bls/src/sig/bls.rs
@@ -158,7 +158,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::curve::zexe::{G1Curve, G2Curve, PairingCurve as PCurve};
+    use crate::curve::bls12377::{G1Curve, G2Curve, PairingCurve as PCurve};
     use crate::group::Curve;
     use rand::prelude::*;
